requirements.txt
asgiref==3.2.7
Django==3.0.6
future==0.18.2
lxml==4.5.1
numpy==1.18.4
Pillow==7.1.2
pytils==0.3
pytz==2020.1
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.1
tools==0.1.9
torch==1.5.0
torchvision==0.6.0
